Moved the AdminState struct from server-primitives
