+++ crates/server/src/admin/handlers/add_client_key.rs
@@ -5,7 +5,7 @@ use axum::response::IntoResponse;
 use axum::{Extension, Json};
 use calimero_primitives::identity::{ClientKey, WalletType};
 use calimero_server_primitives::admin::{
-    AddPublicKeyRequest, AdminState, EthSignatureMessageMetadata, IntermediateAddPublicKeyRequest,
+    AddPublicKeyRequest, EthSignatureMessageMetadata, IntermediateAddPublicKeyRequest,
     JwtRefreshRequest, JwtTokenRequest, NearSignatureMessageMetadata, Payload,
     SignatureMetadataEnum, StarknetSignatureMessageMetadata,
 };
@@ -22,6 +22,7 @@ use crate::admin::storage::client_keys::add_client_key;
 use crate::admin::storage::root_key::exists_root_keys;
 use crate::admin::utils::auth::{validate_challenge, validate_root_key_exists};
 use crate::admin::utils::jwt::{generate_jwt_tokens, refresh_access_token};
+use crate::AdminState;
 
 pub fn transform_request(
     intermediate: IntermediateAddPublicKeyRequest,
