+++ crates/server/src/admin/handlers/context.rs
@@ -7,7 +7,7 @@ use axum::{Extension, Json};
 use calimero_primitives::context::{Context, ContextId};
 use calimero_primitives::identity::{ClientKey, ContextUser};
 use calimero_server_primitives::admin::{
-    AdminState, ContextStorage, CreateContextRequest, CreateContextResponse, GetContextsResponse,
+    ContextStorage, CreateContextRequest, CreateContextResponse, GetContextsResponse,
     UpdateContextApplicationRequest,
 };
 use reqwest::StatusCode;
@@ -18,6 +18,7 @@ use tracing::error;
 use crate::admin::service::{parse_api_error, ApiError, ApiResponse, Empty};
 use crate::admin::storage::client_keys::get_context_client_key;
 use crate::admin::utils::context::{create_context, join_context};
+use crate::AdminState;
 
 #[derive(Debug, Deserialize, Serialize)]
 #[serde(rename_all = "camelCase")]
